moved hedera types to types/ folder and associated import changes
