+++ packages/hedera-wallet-snap/packages/snap/src/services/impl/hedera/client/approveAllowance.ts
@@ -28,7 +28,7 @@ import {
 import { ApproveAllowanceAssetDetail } from '../../../../types/params';
 import { uint8ArrayToHex } from '../../../../utils/crypto';
 import { timestampToString } from '../../../../utils/helper';
-import { TxReceipt } from '../../../hedera';
+import { TxReceipt } from '../../../../types/hedera';
 
 /**
  * Approve an allowance.
