+++ packages/hedera-wallet-snap/packages/snap/src/services/impl/hedera/client/hts/associateTokens.ts
@@ -26,7 +26,7 @@ import {
 
 import { uint8ArrayToHex } from '../../../../../utils/crypto';
 import { timestampToString } from '../../../../../utils/helper';
-import { TxReceipt } from '../../../../hedera';
+import { TxReceipt } from '../../../../../types/hedera';
 
 /**
  * Associate tokens to an account.
